requirements.txt

captum
flask
torch
torchvision
transformers
fpdf2
reportlab
pillow
gunicorn
charset-normalizer